requirements.txt
altgraph==0.17
anaconda==0.0.1.1
attrs==19.2.0
backcall==0.1.0
BareNecessities==0.2.8
beautifulsoup4==4.4.1
bleach==3.1.0
boto3==1.9.139
botocore==1.12.139
cachetools==3.1.1
certifi==2018.4.16
chardet==3.0.4
click==7.1.1
colorama==0.4.1
configparser==5.0.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-storages==1.7.1
docutils==0.14
docx-mailmerge==0.5.0
entrypoints==0.3
et-xmlfile==1.0.1
future==0.18.2
glob3==0.0.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.13.0
idna==2.6
IMAPClient==2.0.0
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
kiwisolver==1.1.0
lxml==4.3.4
Mail==2.1.0
mailchimp3==3.0.9
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.18.4
openpyxl==3.0.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.5.1
pefile==2019.4.18
pep8==1.7.1
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyAutoGUI==0.9.36
Pygments==2.4.2
PyInstaller==3.6
PyJWT==1.6.1
PyMsgBox==1.0.6
pyparsing==2.4.6
PyPDF2==1.26.0
pyperclip==1.6.0
pyrsistent==0.15.4
PyScreeze==0.1.14
PySocks==1.6.8
python-dateutil==2.8.1
python-docx==0.8.10
PyTweening==1.0.3
pytz==2020.1
pywin32==225
pywin32-ctypes==0.2.0
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.10.0
rsa==4.0
s3transfer==0.2.0
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
sh==1.12.14
six==1.14.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
twilio==6.12.0
uritemplate==3.0.0
urllib3==1.24.2
virtualenv==16.6.1
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1.4
widgetsnbextension==3.5.1
zipp==0.6.0
